move compat to bokeh.core
