+++ bokeh/core/compat/mpl.py
