+++ bokeh/core/compat/mplexporter/_py3k_compat.py
