+++ bokeh/core/compat/mplexporter/exporter.py
