+++ bokeh/core/compat/mplexporter/renderers/__init__.py
