+++ bokeh/core/compat/mplexporter/renderers/base.py
