+++ bokeh/core/compat/mplexporter/renderers/fake_renderer.py
