+++ bokeh/core/compat/mplexporter/renderers/vega_renderer.py
