+++ bokeh/core/compat/mplexporter/renderers/vincent_renderer.py
