+++ bokeh/core/compat/mplexporter/tools.py
