+++ bokeh/core/compat/mplexporter/utils.py
