+++ bokeh/mpl.py
@@ -10,4 +10,4 @@ Supporting objects and functions to convert Matplotlib objects into Bokeh.
 
 """
 
-from .compat.mpl import  to_bokeh; to_bokeh
+from .core.compat.mpl import  to_bokeh; to_bokeh
